sms variable cleanup

+++ src/burn/drv/sms/smsrender.cpp
@@ -3,36 +3,35 @@
     Display rendering.
 */
 
-#include "tiles_generic.h"
 #include "smsshared.h"
 
-uint8 sms_cram_expand_table[4];
-uint8 gg_cram_expand_table[16];
+UINT8 sms_cram_expand_table[4];
+UINT8 gg_cram_expand_table[16];
 
 /* Background drawing function */
-void (*render_bg)(int line) = NULL;
-void (*render_obj)(int line) = NULL;
+void (*render_bg)(INT16 line) = NULL;
+void (*render_obj)(INT16 line) = NULL;
 
 /* Pointer to output buffer */
-uint8 *linebuf;
+UINT8 *linebuf;
 
 /* Internal buffer for drawing non 8-bit displays */
-uint8 internal_buffer[0x200];
+UINT8 internal_buffer[0x200];
 
 /* Precalculated pixel table */
-uint16 pixel[PALETTE_SIZE];
+UINT16 pixel[PALETTE_SIZE];
 
 /* Dirty pattern info */
-uint8 bg_name_dirty[0x200];     /* 1= This pattern is dirty */
-uint16 bg_name_list[0x200];     /* List of modified pattern indices */
-uint16 bg_list_index;           /* # of modified patterns in list */
-uint8 bg_pattern_cache[0x20000];/* Cached and flipped patterns */
+UINT8 bg_name_dirty[0x200];     /* 1= This pattern is dirty */
+UINT16 bg_name_list[0x200];     /* List of modified pattern indices */
+UINT16 bg_list_index;           /* # of modified patterns in list */
+UINT8 bg_pattern_cache[0x20000];/* Cached and flipped patterns */
 
 /* Pixel look-up table */
-uint8 lut[0x10000];
+UINT8 lut[0x10000];
 
 /* Attribute expansion table */
-static const uint32 atex[4] =
+static const UINT32 atex[4] =
 {
     0x00000000,
     0x10101010,
@@ -41,58 +40,58 @@ static const uint32 atex[4] =
 };
 
 /* Bitplane to packed pixel LUT */
-uint32 bp_lut[0x10000];
+UINT32 bp_lut[0x10000];
 
-uint32 gg_overscanmode;
+UINT32 gg_overscanmode;
 
 /* Macros to access memory 32-bits at a time (from MAME's drawgfx.c) */
 
 #ifdef ALIGN_DWORD
 
-static __inline__ uint32 read_dword(void *address)
+static __inline__ UINT32 read_dword(void *address)
 {
-    if ((uint32)address & 3)
+    if ((UINT32)address & 3)
 	{
 #ifdef LSB_FIRST  /* little endian version */
-        return ( *((uint8 *)address) +
-                (*((uint8 *)address+1) << 8)  +
-                (*((uint8 *)address+2) << 16) +
-                (*((uint8 *)address+3) << 24) );
+        return ( *((UINT8 *)address) +
+                (*((UINT8 *)address+1) << 8)  +
+                (*((UINT8 *)address+2) << 16) +
+                (*((UINT8 *)address+3) << 24) );
 #else             /* big endian version */
-        return ( *((uint8 *)address+3) +
-                (*((uint8 *)address+2) << 8)  +
-                (*((uint8 *)address+1) << 16) +
-                (*((uint8 *)address)   << 24) );
+        return ( *((UINT8 *)address+3) +
+                (*((UINT8 *)address+2) << 8)  +
+                (*((UINT8 *)address+1) << 16) +
+                (*((UINT8 *)address)   << 24) );
 #endif
 	}
 	else
-        return *(uint32 *)address;
+        return *(UINT32 *)address;
 }
 
 
-static __inline__ void write_dword(void *address, uint32 data)
+static __inline__ void write_dword(void *address, UINT32 data)
 {
-    if ((uint32)address & 3)
+    if ((UINT32)address & 3)
 	{
 #ifdef LSB_FIRST
-            *((uint8 *)address) =    data;
-            *((uint8 *)address+1) = (data >> 8);
-            *((uint8 *)address+2) = (data >> 16);
-            *((uint8 *)address+3) = (data >> 24);
+            *((UINT8 *)address) =    data;
+            *((UINT8 *)address+1) = (data >> 8);
+            *((UINT8 *)address+2) = (data >> 16);
+            *((UINT8 *)address+3) = (data >> 24);
 #else
-            *((uint8 *)address+3) =  data;
-            *((uint8 *)address+2) = (data >> 8);
-            *((uint8 *)address+1) = (data >> 16);
-            *((uint8 *)address)   = (data >> 24);
+            *((UINT8 *)address+3) =  data;
+            *((UINT8 *)address+2) = (data >> 8);
+            *((UINT8 *)address+1) = (data >> 16);
+            *((UINT8 *)address)   = (data >> 24);
 #endif
 		return;
   	}
   	else
-        *(uint32 *)address = data;
+        *(UINT32 *)address = data;
 }
 #else
-#define read_dword(address) *(uint32 *)address
-#define write_dword(address,data) *(uint32 *)address=data
+#define read_dword(address) *(UINT32 *)address
+#define write_dword(address,data) *(UINT32 *)address=data
 #endif
 
 
@@ -106,8 +105,8 @@ void render_shutdown(void)
 /* Initialize the rendering data */
 void render_init(void)
 {
-    int i, j;
-    int bx, sx, b, s, bp, bf, sf, c;
+    INT32 i, j;
+    INT32 bx, sx, b, s, bp, bf, sf, c;
 
     make_tms_tables();
 	memset(&lut, 0, sizeof(lut));
@@ -185,12 +184,12 @@ void render_init(void)
     for(i = 0; i < 0x100; i++)
     for(j = 0; j < 0x100; j++)
     {
-        int x;
-        uint32 out = 0;
+        INT32 x;
+        UINT32 out = 0;
         for(x = 0; x < 8; x++)
         {
-            out |= (j & (0x80 >> x)) ? (uint32)(8 << (x << 2)) : 0;
-            out |= (i & (0x80 >> x)) ? (uint32)(4 << (x << 2)) : 0;
+            out |= (j & (0x80 >> x)) ? (UINT32)(8 << (x << 2)) : 0;
+            out |= (i & (0x80 >> x)) ? (UINT32)(4 << (x << 2)) : 0;
         }
 #if LSB_FIRST
         bp_lut[(j << 8) | (i)] = out;
@@ -201,13 +200,13 @@ void render_init(void)
 
     for(i = 0; i < 4; i++)
     {
-        uint8 c2 = i << 6 | i << 4 | i << 2 | i;
+        UINT8 c2 = i << 6 | i << 4 | i << 2 | i;
         sms_cram_expand_table[i] = c2;
     }
 
     for(i = 0; i < 16; i++)
     {
-        uint8 c2 = i << 4 | i;
+        UINT8 c2 = i << 4 | i;
         gg_cram_expand_table[i] = c2;
     }
 
@@ -219,7 +218,7 @@ void render_init(void)
 /* Reset the rendering data */
 void render_reset(void)
 {
-    int i;
+    INT32 i;
 
     /* Clear display bitmap */
     memset(bitmap.data, 0, bitmap.width * bitmap.height);
@@ -243,7 +242,7 @@ void render_reset(void)
 
 
 /* Draw a line of the display */
-void render_line(int line)
+void render_line(INT16 line)
 {
     /* Ensure we're within the viewport range */
     if(line >= vdp.height)
@@ -298,26 +297,26 @@ void render_line(int line)
 
 
 /* Draw the Master System background */
-void render_bg_sms(int line)
+void render_bg_sms(INT16 line)
 {
-    int locked = 0;
-    int yscroll_mask = (vdp.extended) ? 256 : 224;
-    int v_line = (line + vdp.reg[9]) % yscroll_mask;
-    int v_row  = (v_line & 7) << 3;
-    int hscroll = ((vdp.reg[0] & 0x40) && (line < 0x10)) ? 0 : (0x100 - vdp.reg[8]);
-    int column = 0;
-    uint16 attr;
-    uint16 *nt = (uint16 *)&vdp.vram[vdp.ntab + ((v_line >> 3) << 6)];
-    int nt_scroll = (hscroll >> 3);
-    int shift = (hscroll & 7);
-    uint32 atex_mask;
-    uint32 *cache_ptr;
-    uint32 *linebuf_ptr = (uint32 *)&linebuf[0 - shift];
+    INT32 locked = 0;
+    INT32 yscroll_mask = (vdp.extended) ? 256 : 224;
+    INT32 v_line = (line + vdp.reg[9]) % yscroll_mask;
+    INT32 v_row  = (v_line & 7) << 3;
+    INT32 hscroll = ((vdp.reg[0] & 0x40) && (line < 0x10)) ? 0 : (0x100 - vdp.reg[8]);
+    INT32 column = 0;
+    UINT16 attr;
+    UINT16 *nt = (UINT16 *)&vdp.vram[vdp.ntab + ((v_line >> 3) << 6)];
+    INT32 nt_scroll = (hscroll >> 3);
+    INT32 shift = (hscroll & 7);
+    UINT32 atex_mask;
+    UINT32 *cache_ptr;
+    UINT32 *linebuf_ptr = (UINT32 *)&linebuf[0 - shift];
 
     /* Draw first column (clipped) */
     if(shift)
     {
-        int x;
+        INT32 x;
 
         for(x = shift; x < 8; x++)
             linebuf[(0 - shift) + (x)] = 0;
@@ -333,7 +332,7 @@ void render_bg_sms(int line)
         {
             locked = 1;
             v_row = (line & 7) << 3;
-            nt = (uint16 *)&vdp.vram[((vdp.reg[2] << 10) & 0x3800) + ((line >> 3) << 6)];
+            nt = (UINT16 *)&vdp.vram[((vdp.reg[2] << 10) & 0x3800) + ((line >> 3) << 6)];
         }
 
         /* Get name table attribute word */
@@ -346,7 +345,7 @@ void render_bg_sms(int line)
         atex_mask = atex[(attr >> 11) & 3];
 
         /* Point to a line of pattern data in cache */
-        cache_ptr = (uint32 *)&bg_pattern_cache[((attr & 0x7FF) << 6) | (v_row)];
+        cache_ptr = (UINT32 *)&bg_pattern_cache[((attr & 0x7FF) << 6) | (v_row)];
         
         /* Copy the left half, adding the attribute bits in */
         write_dword( &linebuf_ptr[(column << 1)] , read_dword( &cache_ptr[0] ) | (atex_mask));
@@ -358,9 +357,9 @@ void render_bg_sms(int line)
     /* Draw last column (clipped) */
     if(shift)
     {
-        int x, c, a;
+        INT32 x, c, a;
 
-        uint8 *p = &linebuf[(0 - shift)+(column << 3)];
+        UINT8 *p = &linebuf[(0 - shift)+(column << 3)];
 
         attr = nt[(column + nt_scroll) & 0x1F];
 
@@ -377,24 +376,21 @@ void render_bg_sms(int line)
     }
 }
 
-
-
-
 /* Draw sprites */
-void render_obj_sms(int line)
+void render_obj_sms(INT16 line)
 {
-    int i;
-    uint8 collision_buffer = 0;
+    INT32 i;
+    UINT8 collision_buffer = 0;
 
     /* Sprite count for current line (8 max.) */
-    int count = 0;
+    INT32 count = 0;
 
     /* Sprite dimensions */
-    int width = 8;
-    int height = (vdp.reg[1] & 0x02) ? 16 : 8;
+    INT32 width = 8;
+    INT32 height = (vdp.reg[1] & 0x02) ? 16 : 8;
 
     /* Pointer to sprite attribute table */
-    uint8 *st = (uint8 *)&vdp.vram[vdp.satb];
+    UINT8 *st = (UINT8 *)&vdp.vram[vdp.satb];
 
     /* Adjust dimensions for double size sprites */
     if(vdp.reg[1] & 0x01)
@@ -407,7 +403,7 @@ void render_obj_sms(int line)
     for(i = 0; i < 64; i++)
     {
         /* Sprite Y position */
-        int yp = st[i];
+        INT32 yp = st[i];
 
         /* Found end of sprite list marker for non-extended modes? */
         if(vdp.extended == 0 && yp == 208)
@@ -422,17 +418,17 @@ void render_obj_sms(int line)
         /* Check if sprite falls on current line */
         if((line >= yp) && (line < (yp + height)))
         {
-            uint8 *linebuf_ptr;
+            UINT8 *linebuf_ptr;
 
             /* Width of sprite */
-            int start = 0;
-            int end = width;
+            INT32 start = 0;
+            INT32 end = width;
 
             /* Sprite X position */
-            int xp = st[0x80 + (i << 1)];
+            INT32 xp = st[0x80 + (i << 1)];
 
             /* Pattern name */
-            int n = st[0x81 + (i << 1)];
+            INT32 n = st[0x81 + (i << 1)];
 
             /* Bump sprite count */
             count++;
@@ -454,7 +450,7 @@ void render_obj_sms(int line)
             if(vdp.reg[1] & 0x02) n &= 0x01FE;
 
             /* Point to offset in line buffer */
-            linebuf_ptr = (uint8 *)&linebuf[xp];
+            linebuf_ptr = (UINT8 *)&linebuf[xp];
 
             /* Clip sprites on left edge */
             if(xp < 0)
@@ -471,20 +467,20 @@ void render_obj_sms(int line)
             /* Draw double size sprite */
             if(vdp.reg[1] & 0x01)
             {
-                int x;
-                uint8 *cache_ptr = (uint8 *)&bg_pattern_cache[(n << 6) | (((line - yp) >> 1) << 3)];
+                INT16 x;
+                UINT8 *cache_ptr = (UINT8 *)&bg_pattern_cache[(n << 6) | (((line - yp) >> 1) << 3)];
 
                 /* Draw sprite line */
                 for(x = start; x < end; x++)
                 {
                     /* Source pixel from cache */
-                    uint8 sp = cache_ptr[(x >> 1)];
+                    UINT8 sp = cache_ptr[(x >> 1)];
     
                     /* Only draw opaque sprite pixels */
                     if(sp)
                     {
                         /* Background pixel from line buffer */
-                        uint8 bg = linebuf_ptr[x];
+                        UINT8 bg = linebuf_ptr[x];
     
                         /* Look up result */
                         linebuf_ptr[x] = lut[(bg << 8) | (sp)];
@@ -496,20 +492,20 @@ void render_obj_sms(int line)
             }
             else /* Regular size sprite (8x8 / 8x16) */
             {
-                int x;
-                uint8 *cache_ptr = (uint8 *)&bg_pattern_cache[(n << 6) | ((line - yp) << 3)];
+                INT16 x;
+                UINT8 *cache_ptr = (UINT8 *)&bg_pattern_cache[(n << 6) | ((line - yp) << 3)];
 
                 /* Draw sprite line */
                 for(x = start; x < end; x++)
                 {
                     /* Source pixel from cache */
-                    uint8 sp = cache_ptr[x];
+                    UINT8 sp = cache_ptr[x];
     
                     /* Only draw opaque sprite pixels */
                     if(sp)
                     {
                         /* Background pixel from line buffer */
-                        uint8 bg = linebuf_ptr[x];
+                        UINT8 bg = linebuf_ptr[x];
     
                         /* Look up result */
                         linebuf_ptr[x] = lut[(bg << 8) | (sp)];
@@ -531,9 +527,9 @@ end:
 
 void update_bg_pattern_cache(void)
 {
-    int i;
-    uint8 x, y;
-    uint16 name;
+    INT32 i;
+    UINT8 x, y;
+    UINT16 name;
 
     if(!bg_list_index) return;
 
@@ -546,15 +542,15 @@ void update_bg_pattern_cache(void)
         {
             if(bg_name_dirty[name] & (1 << y))
             {
-                uint8 *dst = &bg_pattern_cache[name << 6];
+                UINT8 *dst = &bg_pattern_cache[name << 6];
 
-                uint16 bp01 = *(uint16 *)&vdp.vram[(name << 5) | (y << 2) | (0)];
-                uint16 bp23 = *(uint16 *)&vdp.vram[(name << 5) | (y << 2) | (2)];
-                uint32 temp = (bp_lut[bp01] >> 2) | (bp_lut[bp23]);
+                UINT16 bp01 = *(UINT16 *)&vdp.vram[(name << 5) | (y << 2) | (0)];
+                UINT16 bp23 = *(UINT16 *)&vdp.vram[(name << 5) | (y << 2) | (2)];
+                UINT32 temp = (bp_lut[bp01] >> 2) | (bp_lut[bp23]);
 
                 for(x = 0; x < 8; x++)
                 {
-                    uint8 c = (temp >> (x << 2)) & 0x0F;
+                    UINT8 c = (temp >> (x << 2)) & 0x0F;
                     dst[0x00000 | (y << 3) | (x)] = (c);
                     dst[0x08000 | (y << 3) | (x ^ 7)] = (c);
                     dst[0x10000 | ((y ^ 7) << 3) | (x)] = (c);
@@ -569,9 +565,9 @@ void update_bg_pattern_cache(void)
 
 
 /* Update a palette entry */
-void palette_sync(int index, int force)
+void palette_sync(INT16 index, INT16 force)
 {
-    int r, g, b;
+    INT16 r, g, b;
 
     // unless we are forcing an update,
     // if not in mode 4, exit
@@ -612,11 +608,11 @@ void palette_sync(int index, int force)
     bitmap.pal.dirty[index] = bitmap.pal.update = 1;
 }
 
-void remap_8_to_16(int line, int extend)
+void remap_8_to_16(INT16 line, INT16 extend)
 {
 	if (line > nScreenHeight || (line - extend) < 0) return;
 
-	UINT16 *p = (uint16 *)&bitmap.data[((line - extend) * bitmap.pitch)];
+	UINT16 *p = (UINT16 *)&bitmap.data[((line - extend) * bitmap.pitch)];
     for (INT32 i = bitmap.viewport.x; i < bitmap.viewport.w + bitmap.viewport.x; i++)
     {
 		p[i] = internal_buffer[i] & PIXEL_MASK;
