+++ src/burn/drv/sms/smstms.cpp
@@ -4,35 +4,35 @@
 */
 #include "smsshared.h"
 
-int text_counter;               /* Text offset counter */
-uint8 tms_lookup[16][256][2];   /* Expand BD, PG data into 8-bit pixels (G1,G2) */
-uint8 mc_lookup[16][256][8];    /* Expand BD, PG data into 8-bit pixels (MC) */
-uint8 txt_lookup[256][2];       /* Expand BD, PG data into 8-bit pixels (TX) */
-uint8 bp_expand[256][8];        /* Expand PG data into 8-bit pixels */
-uint8 tms_obj_lut[16*256];      /* Look up priority between SG and display pixels */
+INT16 text_counter;               /* Text offset counter */
+UINT8 tms_lookup[16][256][2];   /* Expand BD, PG data into 8-bit pixels (G1,G2) */
+UINT8 mc_lookup[16][256][8];    /* Expand BD, PG data into 8-bit pixels (MC) */
+UINT8 txt_lookup[256][2];       /* Expand BD, PG data into 8-bit pixels (TX) */
+UINT8 bp_expand[256][8];        /* Expand PG data into 8-bit pixels */
+UINT8 tms_obj_lut[16*256];      /* Look up priority between SG and display pixels */
 
-static const uint8 diff_mask[]  = {0x07, 0x07, 0x0F, 0x0F};
-static const uint8 name_mask[]  = {0xFF, 0xFF, 0xFC, 0xFC};
-static const uint8 diff_shift[] = {0, 1, 0, 1};
-static const uint8 size_tab[]   = {8, 16, 16, 32};
+static const UINT8 diff_mask[]  = {0x07, 0x07, 0x0F, 0x0F};
+static const UINT8 name_mask[]  = {0xFF, 0xFF, 0xFC, 0xFC};
+static const UINT8 diff_shift[] = {0, 1, 0, 1};
+static const UINT8 size_tab[]   = {8, 16, 16, 32};
 
 /* Internally latched sprite data in the VDP */
 typedef struct {
-    int xpos;
-    uint8 attr;
-    uint8 sg[2];
+    INT16 xpos;
+    UINT8 attr;
+    UINT8 sg[2];
 } tms_sprite;
 
 tms_sprite sprites[4];
-int sprites_found;
+INT16 sprites_found;
 
-void parse_line(int line)
+void parse_line(INT16 line)
 {
-    int yp, i;
-    int mode = vdp.reg[1] & 3;
-    int size = size_tab[mode];
-    int diff, name;
-    uint8 *sa, *sg;
+    INT16 yp, i;
+    INT16 mode = vdp.reg[1] & 3;
+    INT16 size = size_tab[mode];
+    INT16 diff, name;
+    UINT8 *sa, *sg;
     tms_sprite *p;
 
     /* Reset # of sprites found */
@@ -41,7 +41,7 @@ void parse_line(int line)
     /* Parse sprites */
     for(i = 0; i < 32; i++)
     {
-        /* Point to current sprite in SA and our current sprite record */
+        /* PoINT16 to current sprite in SA and our current sprite record */
         p = &sprites[sprites_found];
         sa = &vdp.vram[vdp.sa + (i << 2)];
 
@@ -102,11 +102,11 @@ parse_end:
     vdp.status = (vdp.status & 0xE0) | (i & 0x1F);
 }
 
-void render_obj_tms(int /*line*/)
+void render_obj_tms(INT16 /*line*/)
 {
-    int i, x = 0;
-    int size, start, end, mode;
-    uint8 *lb, *lutp, *ex[2];
+    INT16 i, x = 0;
+    INT16 size, start, end, mode;
+    UINT8 *lb, *lutp, *ex[2];
     tms_sprite *p;
 
     mode = vdp.reg[1] & 3;
@@ -119,7 +119,7 @@ void render_obj_tms(int /*line*/)
         lb = &linebuf[p->xpos];
         lutp = &tms_obj_lut[(p->attr & 0x0F) << 8];
 
-        /* Point to expanded PG data */
+        /* PoINT16 to expanded PG data */
         ex[0] = bp_expand[p->sg[0]];
         ex[1] = bp_expand[p->sg[1]];
 
@@ -227,19 +227,19 @@ void render_obj_tms(int /*line*/)
 
 void make_tms_tables(void)
 {
-    int i, j, x;
-    int bd, pg, ct;
-    int sx, bx;
+    INT16 i, j, x;
+    INT16 bd, pg, ct;
+    INT16 sx, bx;
 
     for(sx = 0; sx < 16; sx++)
     {
         for(bx = 0; bx < 256; bx++)
         {
-//          uint8 bd = (bx & 0x0F);
-            uint8 bs = (bx & 0x40);
-//          uint8 bt = (bd == 0) ? 1 : 0;
-            uint8 sd = (sx & 0x0F);
-//          uint8 st = (sd == 0) ? 1 : 0;
+//          UINT8 bd = (bx & 0x0F);
+            UINT8 bs = (bx & 0x40);
+//          UINT8 bt = (bd == 0) ? 1 : 0;
+            UINT8 sd = (sx & 0x0F);
+//          UINT8 st = (sd == 0) ? 1 : 0;
 
             // opaque sprite pixel, choose 2nd pal and set sprite marker
             if(sd && !bs)
@@ -263,8 +263,8 @@ void make_tms_tables(void)
     /* Text lookup table */
     for(bd = 0; bd < 256; bd++)
     {
-        uint8 bg = (bd >> 0) & 0x0F;
-        uint8 fg = (bd >> 4) & 0x0F;
+        UINT8 bg = (bd >> 0) & 0x0F;
+        UINT8 fg = (bd >> 4) & 0x0F;
 
         /* If foreground is transparent, use background color */
         if(fg == 0) fg = bg;
@@ -278,8 +278,8 @@ void make_tms_tables(void)
     {
         for(pg = 0; pg < 256; pg++)
         {
-            int l = (pg >> 4) & 0x0F;
-            int r = (pg >> 0) & 0x0F;
+            INT16 l = (pg >> 4) & 0x0F;
+            INT16 r = (pg >> 0) & 0x0F;
 
             /* If foreground is transparent, use background color */
             if(l == 0) l = bd;
@@ -288,7 +288,7 @@ void make_tms_tables(void)
             /* Unpack 2 nibbles across eight pixels */
             for(x = 0; x < 8; x++)
             {
-                int c = (x & 4) ? r : l;
+                INT16 c = (x & 4) ? r : l;
 
                 mc_lookup[bd][pg][x] = c;
             }
@@ -301,7 +301,7 @@ void make_tms_tables(void)
     {
         for(j = 0; j < 8; j++)
         {
-            int c = (i >> (j ^ 7)) & 1;
+            INT16 c = (i >> (j ^ 7)) & 1;
             bp_expand[i][j] = c;
         }
     }
@@ -311,9 +311,9 @@ void make_tms_tables(void)
     {
         for(ct = 0; ct < 0x100; ct++)
         {
-            int backdrop = (bd & 0x0F);
-            int background = (ct >> 0) & 0x0F;
-            int foreground = (ct >> 4) & 0x0F;
+            INT16 backdrop = (bd & 0x0F);
+            INT16 background = (ct >> 0) & 0x0F;
+            INT16 foreground = (ct >> 4) & 0x0F;
 
             /* If foreground is transparent, use background color */
             if(background == 0) background = backdrop;
@@ -326,7 +326,7 @@ void make_tms_tables(void)
 }
 
 
-void render_bg_tms(int line)
+void render_bg_tms(INT16 line)
 {
     switch(vdp.mode & 7)
     {
@@ -365,18 +365,18 @@ void render_bg_tms(int line)
 }
 
 /* Graphics I */
-void render_bg_m0(int line)
+void render_bg_m0(INT16 line)
 {
-    int v_row  = (line & 7);
-    int column;
-    int name;
+    INT16 v_row  = (line & 7);
+    INT16 column;
+    INT16 name;
 
-    uint8 *clut;
-    uint8 *bpex;
-    uint8 *lb = &linebuf[0];
-    uint8 *pn = &vdp.vram[vdp.pn + ((line >> 3) << 5)];
-    uint8 *ct = &vdp.vram[vdp.ct];
-    uint8 *pg = &vdp.vram[vdp.pg | (v_row)];
+    UINT8 *clut;
+    UINT8 *bpex;
+    UINT8 *lb = &linebuf[0];
+    UINT8 *pn = &vdp.vram[vdp.pn + ((line >> 3) << 5)];
+    UINT8 *ct = &vdp.vram[vdp.ct];
+    UINT8 *pg = &vdp.vram[vdp.pg | (v_row)];
 
     for(column = 0; column < 32; column++)
     {
@@ -388,20 +388,20 @@ void render_bg_m0(int line)
 }
 
 /* Text */
-void render_bg_m1(int line)
+void render_bg_m1(INT16 line)
 {
-    int v_row  = (line & 7);
-    int column;
+    INT16 v_row  = (line & 7);
+    INT16 column;
 
-    uint8 *clut;
-    uint8 *bpex;
-    uint8 *lb = &linebuf[0];
-//  uint8 *pn = &vdp.vram[vdp.pn + ((line >> 3) * 40)];
+    UINT8 *clut;
+    UINT8 *bpex;
+    UINT8 *lb = &linebuf[0];
+//  UINT8 *pn = &vdp.vram[vdp.pn + ((line >> 3) * 40)];
 
-    uint8 *pn = &vdp.vram[vdp.pn + text_counter];
+    UINT8 *pn = &vdp.vram[vdp.pn + text_counter];
 
-    uint8 *pg = &vdp.vram[vdp.pg | (v_row)];
-    uint8 bk = vdp.reg[7];
+    UINT8 *pg = &vdp.vram[vdp.pg | (v_row)];
+    UINT8 bk = vdp.reg[7];
 
     clut = &txt_lookup[bk][0];
 
@@ -419,17 +419,17 @@ void render_bg_m1(int line)
 }
 
 /* Text + extended PG */
-void render_bg_m1x(int line)
+void render_bg_m1x(INT16 line)
 {
-    int v_row  = (line & 7);
-    int column;
+    INT16 v_row  = (line & 7);
+    INT16 column;
 
-    uint8 *clut;
-    uint8 *bpex;
-    uint8 *lb = &linebuf[0];
-    uint8 *pn = &vdp.vram[vdp.pn + ((line >> 3) * 40)];
-    uint8 *pg = &vdp.vram[vdp.pg + (v_row) + ((line & 0xC0) << 5)];
-    uint8 bk = vdp.reg[7];
+    UINT8 *clut;
+    UINT8 *bpex;
+    UINT8 *lb = &linebuf[0];
+    UINT8 *pn = &vdp.vram[vdp.pn + ((line >> 3) * 40)];
+    UINT8 *pg = &vdp.vram[vdp.pg + (v_row) + ((line & 0xC0) << 5)];
+    UINT8 bk = vdp.reg[7];
 
     clut = &tms_lookup[0][bk][0];
 
@@ -442,13 +442,13 @@ void render_bg_m1x(int line)
 }
 
 /* Invalid (2+3/1+2+3) */
-void render_bg_inv(int /*line*/)
+void render_bg_inv(INT16 /*line*/)
 {
-    int column;
-    uint8 *clut;
-    uint8 *bpex;
-    uint8 *lb = &linebuf[0];
-    uint8 bk = vdp.reg[7];
+    INT16 column;
+    UINT8 *clut;
+    UINT8 *bpex;
+    UINT8 *lb = &linebuf[0];
+    UINT8 bk = vdp.reg[7];
 
     clut = &txt_lookup[bk][0];
 
@@ -460,14 +460,14 @@ void render_bg_inv(int /*line*/)
 }
 
 /* Multicolor */
-void render_bg_m3(int line)
+void render_bg_m3(INT16 line)
 {
-    int column;
-    uint8 *mcex;
-    uint8 *lb = &linebuf[0];
+    INT16 column;
+    UINT8 *mcex;
+    UINT8 *lb = &linebuf[0];
 
-    uint8 *pn = &vdp.vram[vdp.pn + ((line >> 3) << 5)];
-    uint8 *pg = &vdp.vram[vdp.pg + ((line >> 2) & 7)];
+    UINT8 *pn = &vdp.vram[vdp.pn + ((line >> 3) << 5)];
+    UINT8 *pg = &vdp.vram[vdp.pg + ((line >> 2) & 7)];
 
     for(column = 0; column < 32; column++)
     {
@@ -477,13 +477,13 @@ void render_bg_m3(int line)
 }
 
 /* Multicolor + extended PG */
-void render_bg_m3x(int line)
+void render_bg_m3x(INT16 line)
 {
-    int column;
-    uint8 *mcex;
-    uint8 *lb = &linebuf[0];
-    uint8 *pn = &vdp.vram[vdp.pn + ((line >> 3) << 5)];
-    uint8 *pg = &vdp.vram[vdp.pg + ((line >> 2) & 7) + ((line & 0xC0) << 5)];
+    INT16 column;
+    UINT8 *mcex;
+    UINT8 *lb = &linebuf[0];
+    UINT8 *pn = &vdp.vram[vdp.pn + ((line >> 3) << 5)];
+    UINT8 *pg = &vdp.vram[vdp.pg + ((line >> 2) & 7) + ((line & 0xC0) << 5)];
 
     for(column = 0; column < 32; column++)
     {
@@ -493,18 +493,18 @@ void render_bg_m3x(int line)
 }
 
 /* Graphics II */
-void render_bg_m2(int line)
+void render_bg_m2(INT16 line)
 {
-    int v_row  = (line & 7);
-    int column;
-    int name;
+    INT16 v_row  = (line & 7);
+    INT16 column;
+    INT16 name;
 
-    uint8 *clut;
-    uint8 *bpex;
-    uint8 *lb = &linebuf[0];
-    uint8 *pn = &vdp.vram[vdp.pn | ((line & 0xF8) << 2)];
-    uint8 *ct = &vdp.vram[(vdp.ct & 0x2000) | (v_row) | ((line & 0xC0) << 5)];
-    uint8 *pg = &vdp.vram[(vdp.pg & 0x2000) | (v_row) | ((line & 0xC0) << 5)];
+    UINT8 *clut;
+    UINT8 *bpex;
+    UINT8 *lb = &linebuf[0];
+    UINT8 *pn = &vdp.vram[vdp.pn | ((line & 0xF8) << 2)];
+    UINT8 *ct = &vdp.vram[(vdp.ct & 0x2000) | (v_row) | ((line & 0xC0) << 5)];
+    UINT8 *pg = &vdp.vram[(vdp.pg & 0x2000) | (v_row) | ((line & 0xC0) << 5)];
 
     for(column = 0; column < 32; column++)
     {
